flask_app/requirement.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.1.3
colorama==0.4.4
dateparser==1.1.0
Django==4.0.3
docopt==0.6.2
ephem==4.1.1
Flask==2.1.2
fonttools==4.28.2
hijri-converter==2.2.2
holidays==0.11.3.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pipreqs==0.4.11
prometheus-api-client==0.4.2
prometheus-client==0.12.0
PyMeeus==0.5.11
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
Werkzeug==2.1.2
yarg==0.1.9
